+++ Scripts/Runtime/MenuModal.cs
@@ -2,17 +2,18 @@
 using UnityEngine.UI;
 using TMPro;
 using Vulpes.Promises;
+using Vulpes.Transitions;
 using System;
 
 namespace Vulpes.Menus
 {
     /// <summary>
-    /// The <see cref="MenuDialogue"/> is a verbose popup that can be used to query the user 
+    /// The <see cref="MenuModal"/> is a verbose popup that can be used to query the user 
     /// and can be displayed with either one, two or three configurable buttons. The resulting 
-    /// <see cref="Promise"/> resolves with a <see cref="MenuDialogueResult"/> based on the <see cref="Button"/> pressed.
+    /// <see cref="Promise"/> resolves with a <see cref="MenuModalResult"/> based on the <see cref="Button"/> pressed.
     /// </summary>
-    [AddComponentMenu("Vulpes/Menus/Menu Dialogue")]
-    public class MenuDialogue : MenuScreen, IMenuDialogue
+    [AddComponentMenu("Vulpes/Menus/Menu Modal")]
+    public class MenuModal : MenuScreen, IMenuModal
     {
         [SerializeField] private TextMeshProUGUI titleText = default;
         [SerializeField] private TextMeshProUGUI bodyText = default;
@@ -25,7 +26,7 @@ namespace Vulpes.Menus
         [SerializeField] private Button alternateButton = default;
 
         /// <summary>
-        /// Sets the <see cref="TextMeshProUGUI"/> for the <see cref="MenuDialogue"/> and disables any <see cref="Button"/>s with no text string.
+        /// Sets the <see cref="TextMeshProUGUI"/> for the <see cref="MenuModal"/> and disables any <see cref="Button"/>s with no text string.
         /// </summary>
         private void SetText(in string title, in string body, in string confirm, in string cancel, in string alternate)
         {
@@ -42,44 +43,44 @@ namespace Vulpes.Menus
         }
 
         /// <summary>
-        /// Transitions in a <see cref="MenuDialogue"/> with one <see cref="Button"/> and returns a <see cref="Promise"/> that resolves with a <see cref="MenuDialogueResult"/>
+        /// Transitions in a <see cref="MenuModal"/> with one <see cref="Button"/> and returns a <see cref="Promise"/> that resolves with a <see cref="MenuModalResult"/>
         /// result when the <see cref="MenuTransition"/> is complete (Note: Callbacks execute before the <see cref="Promise"/> resolves).
         /// </summary>
-        public IPromise<MenuDialogueResult> Show(in string title, in string body, in string confirm, Action onConfirm = null)
+        public IPromise<MenuModalResult> Show(in string title, in string body, in string confirm, Action onConfirm = null)
         {
             return ShowInternal(title, body, confirm, null, null, onConfirm, null, null);
         }
 
         /// <summary>
-        /// Transitions in a <see cref="MenuDialogue"/> with two <see cref="Button"/>s and returns a Promise that resolves with a <see cref="MenuDialogueResult"/>
+        /// Transitions in a <see cref="MenuModal"/> with two <see cref="Button"/>s and returns a Promise that resolves with a <see cref="MenuModalResult"/>
         /// result when the <see cref="MenuTransition"/> is complete (Note: Callbacks execute before the <see cref="Promise"/> resolves).
         /// </summary>
-        public IPromise<MenuDialogueResult> Show(in string title, in string body, in string confirm, in string cancel, Action onConfirm = null, Action onCancel = null)
+        public IPromise<MenuModalResult> Show(in string title, in string body, in string confirm, in string cancel, Action onConfirm = null, Action onCancel = null)
         {
             return ShowInternal(title, body, confirm, cancel, null, onConfirm, onCancel, null);
         }
 
         /// <summary>
-        /// Transitions in a <see cref="MenuDialogue"/> with three <see cref="Button"/>s and returns a <see cref="Promise"/> that resolves with a <see cref="MenuDialogueResult"/>
+        /// Transitions in a <see cref="MenuModal"/> with three <see cref="Button"/>s and returns a <see cref="Promise"/> that resolves with a <see cref="MenuModalResult"/>
         /// result when the <see cref="MenuTransition"/> is complete (Note: Callbacks execute before the <see cref="Promise"/> resolves).
         /// </summary>
-        public IPromise<MenuDialogueResult> Show(in string title, in string body, in string confirm, in string cancel, in string alternate, Action onConfirm = null, Action onCancel = null, Action onAlternate = null)
+        public IPromise<MenuModalResult> Show(in string title, in string body, in string confirm, in string cancel, in string alternate, Action onConfirm = null, Action onCancel = null, Action onAlternate = null)
         {
             return ShowInternal(title, body, confirm, cancel, alternate, onConfirm, onCancel, onAlternate);
         }
 
         /// <summary>
-        /// Transitions in the <see cref="MenuDialogue"/> and returns a <see cref="Promise"/> that resolves with a <see cref="MenuDialogueResult"/> when a <see cref="Button"/> is pressed.
+        /// Transitions in the <see cref="MenuModal"/> and returns a <see cref="Promise"/> that resolves with a <see cref="MenuModalResult"/> when a <see cref="Button"/> is pressed.
         /// </summary>
-        private IPromise<MenuDialogueResult> ShowInternal(in string title, in string body, in string confirm, in string cancel, in string alternate, Action onConfirm = null, Action onCancel = null, Action onAlternate = null)
+        private IPromise<MenuModalResult> ShowInternal(in string title, in string body, in string confirm, in string cancel, in string alternate, Action onConfirm = null, Action onCancel = null, Action onAlternate = null)
         {
             if (IsCurrentScreen)
             {
                 Debug.LogWarning("Menu Dialogue is already in use, you cannot display another Dialogue until the existing one is dismissed.");
-                return Promise<MenuDialogueResult>.Rejected(null);
+                return Promise<MenuModalResult>.Rejected(null);
             }
 
-            IPromise<MenuDialogueResult> promise = Promise<MenuDialogueResult>.Create();
+            IPromise<MenuModalResult> promise = Promise<MenuModalResult>.Create();
 
             SetText(title, body, confirm, cancel, alternate);
 
@@ -91,21 +92,21 @@ namespace Vulpes.Menus
             {
                 HideInternal().Done(() => {
                     onConfirm?.Invoke();
-                    promise.Resolve(MenuDialogueResult.Confirm);
+                    promise.Resolve(MenuModalResult.Confirm);
                 });
             });
             cancelButton.onClick.AddListener(() =>
             {
                 HideInternal().Done(() => {
                     onCancel?.Invoke();
-                    promise.Resolve(MenuDialogueResult.Cancel);
+                    promise.Resolve(MenuModalResult.Cancel);
                 });
             });
             alternateButton.onClick.AddListener(() =>
             {
                 HideInternal().Done(() => {
                     onAlternate?.Invoke();
-                    promise.Resolve(MenuDialogueResult.Alternate);
+                    promise.Resolve(MenuModalResult.Alternate);
                 });
             });
 
@@ -115,7 +116,7 @@ namespace Vulpes.Menus
         }
 
         /// <summary>
-        /// Transitions out the <see cref="MenuDialogue"/> and returns a <see cref="Promise"/> that resolves when the <see cref="MenuTransition"/> is complete.
+        /// Transitions out the <see cref="MenuModal"/> and returns a <see cref="Promise"/> that resolves when the <see cref="Transition"/> is complete.
         /// </summary>
         private IPromise HideInternal()
         {